requirements.txt
python-telegram-bot==20.8
APScheduler==3.10.4
pytz==2023.3.post1
nest_asyncio==1.5.8
flask
